tests/wpt/tests/js-self-profiling/with-document-policy/dedicated-worker-blob.https.html

<!DOCTYPE html>
<html>
<head>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
  <script>
    // A blob: URL worker should inherit Document-Policy from the creator
    // document, allowing profiling when the creator has js-profiling enabled.
    promise_test(async t => {
      const workerScript = `
        try {
          new Profiler({ sampleInterval: 10, maxBufferSize: Number.MAX_SAFE_INTEGER });
          self.postMessage({ success: true, message: 'Profiler created' });
        } catch (e) {
          self.postMessage({ success: false, error: e.name, message: e.message });
        }
      `;
      const blob = new Blob([workerScript], { type: 'text/javascript' });
      const url = URL.createObjectURL(blob);
      const worker = new Worker(url);
      const result = await new Promise((resolve, reject) => {
        worker.onmessage = e => resolve(e.data);
        worker.onerror = e => reject(new Error(e.message));
      });
      assert_true(result.success,
                  'profiling should succeed in blob: dedicated worker with inherited document policy');
      worker.terminate();
      URL.revokeObjectURL(url);
    }, 'blob: dedicated worker should inherit js-profiling policy from creator document');
  </script>
</body>
</html>
